Update session_key <-> access_token exchange
The new version of HTTParty (perhaps because of multi_json?) is wrapping a 1-key
response in an array where before it wasn't. This updates the specs to match and
makes the mocked data look more like the `HTTParty::Response` it's emulating.

I've also updated `#get_access_token_for_session_key` to use post without the
`Mogli::Client` instance, since that's what Mike's preferred style seems to be
in `#get_access_token_for_application` and the client offers nothing as a
middleman.

Future work:
The `Mogli::Client#api_path` instance method should probably be a class method,
as it doesn't rely on the state of an instance.