Fix response parsing when receiving 400 on session invalidation.
Couple of fixes required here.  When a session is invalidated (perhaps because
the user changed their password) and we try to post to a user's feed,
we get back an http response object with a 400 error code.  This was
causing a general HTTPException to be raised instead of a session invalidation
exception because we considered anything without a response code of 200 to be
an error.  On 400, we want to parse the response and examine the error
message.