Mogli client is more resilient to Facebook's JSON
Facebook APIs return numbers, "true" and "false" as valid JSON, even though the
JSON RFC defines JSON objects as starting with "[" or "{". As such, when the
client is looking at parsed JSON returned from a Facebook API call, it will
crash on numbers and booleans because they don't respond to the :has_keys?
method.